Use a standard algorithm to find updates
index.txt represents the current set of files and their md5sums.
local_index.txt represents the set of files we currently have on disk
and their md5sums. So, the list of files we need to download is simply
the set subtraction, index - local_index. comm is the standard POSIX
utility that implements this operation.