Install .desktop in $HOME/.local
This reduces the amount of sudo needed and the amount of clutter
outside the user's home directory. This also fixes a failure when the
user asks for a desktop shortcut but no menu shortcut This also fixes a
failure when the user asks for a desktop shortcut but no menu shortcut.